got modal off the ground, still a lot of work to do
