+++ src/components/WeekPage/WeekPage.jsx
@@ -11,11 +11,24 @@ import FormControl from "@mui/material/FormControl";
 import Box from "@mui/material/Box";
 import Grid from "@mui/material/Grid";
 import Stack from "@mui/material/Stack";
+import AddActivityForm from "../AddActivityForm/AddActivityForm";
+import ActivityModal from "../ActivityModal/ActivityModal";
 function WeekPage() {
   // const [activities, setActivities] = useState([]);
   const dispatch = useDispatch();
   const activities = useSelector((store) => store.activities);
+  const [modalOpen, setModalOpen] = useState(false);
+  const [selectedActivity, setSelectedActivity] = useState({});
 
+  const handleActivityClick = (activity) => {
+    setSelectedActivity(activity);
+    setModalOpen(true);
+  };
+
+  const handleCloseModal = () => {
+    setSelectedActivity({});
+    setModalOpen(false);
+  };
   const handleAddActivity = (activity) => {
     const start = new Date(`1970-01-01T${activity.start_time}:00.000Z`);
     const end = new Date(`1970-01-01T${activity.end_time}:00.000Z`);
@@ -23,10 +36,7 @@ function WeekPage() {
     const activityWithTime = { ...activity, total_hours: timeDiff };
     dispatch({ type: "POST_ACTIVITY", payload: activityWithTime });
   };
-  const handleClick = (activity) => {
-    // event.preventDefault();
-    console.log("one of the things got clicked", activity);
-  };
+
   // dispatch({ type: "POST_ACTIVITY", payload: activity });
   // setActivities([...activities, activity]);
 
@@ -70,7 +80,10 @@ function WeekPage() {
                 {activities
                   .filter((activity) => activity.day === day)
                   .map((activity) => (
-                    <li key={activity.id} onClick={() => handleClick(activity)}>
+                    <li
+                      key={activity.id}
+                      onClick={() => handleActivityClick(activity)}
+                    >
                       <Typography>{activity.category_name}</Typography>
                       <Typography>
                         {activity.start_time} - {activity.end_time}
@@ -78,6 +91,11 @@ function WeekPage() {
                     </li>
                   ))}
               </ul>
+              <ActivityModal
+                activity={selectedActivity}
+                open={modalOpen}
+                onClose={handleCloseModal}
+              />
             </FormControl>
           </Stack>
         ))}
@@ -86,134 +104,4 @@ function WeekPage() {
   );
 }
 
-function AddActivityForm({ onAddActivity, activities, daysOfWeek }) {
-  const [activity, setActivity] = useState({
-    category_id: 0,
-    day: "",
-    start_time: "",
-    end_time: "",
-    total_hours: "",
-  });
-  const [overlapError, setOverlapError] = useState(false);
-
-  const handleSubmit = (event) => {
-    event.preventDefault();
-
-    const existingActivities = activities.filter(
-      (a) =>
-        a.day === activity.day &&
-        a !== activity &&
-        ((a.start_time >= activity.start_time &&
-          a.start_time < activity.end_time) ||
-          (activity.start_time >= a.start_time &&
-            activity.start_time < a.end_time))
-    );
-
-    if (existingActivities.length > 0) {
-      setOverlapError(true);
-    } else {
-      onAddActivity({ ...activity });
-      setActivity({
-        category_id: 0,
-        day: "",
-        start_time: "",
-        end_time: "",
-        total_hours: "",
-      });
-      setOverlapError(false);
-    }
-  };
-
-  const handleChange = (event) => {
-    setActivity({ ...activity, [event.target.name]: event.target.value });
-  };
-
-  return (
-    <form onSubmit={handleSubmit}>
-      <Typography variant="h5" gutterBottom>
-        Add an Activity
-      </Typography>
-      <Grid container spacing={1}>
-        <Grid item xs={12} md={2}>
-          <FormControl fullWidth>
-            <Typography>Select Category</Typography>
-            <Select
-              name="category_id"
-              value={activity.category_id}
-              onChange={handleChange}
-              // className={classes.input}
-            >
-              <MenuItem value={0}></MenuItem>
-              <MenuItem value={1}>Sleep</MenuItem>
-              <MenuItem value={2}>Self-Care</MenuItem>
-              <MenuItem value={3}>Family and Relationships</MenuItem>
-              <MenuItem value={4}>Personal Development</MenuItem>
-              <MenuItem value={5}>Nutrition</MenuItem>
-              <MenuItem value={6}>Leisure Time</MenuItem>
-              <MenuItem value={7}>Community Involvement</MenuItem>
-              <MenuItem value={8}>Creativity</MenuItem>
-              <MenuItem value={9}>Work</MenuItem>
-              <MenuItem value={10}>Measure What Matters</MenuItem>
-            </Select>
-          </FormControl>
-        </Grid>
-        <Grid item xs={12} md={2}>
-          <FormControl fullWidth>
-            <Typography>Select Day</Typography>
-            <Select
-              name="day"
-              value={activity.day}
-              onChange={handleChange}
-              // label="Select Day"
-              // className={classes.input}
-            >
-              <MenuItem value="0">Select Day</MenuItem>
-              {daysOfWeek.map((day) => (
-                <MenuItem key={day} value={day}>
-                  {day}
-                </MenuItem>
-              ))}
-            </Select>
-          </FormControl>
-        </Grid>
-        <Grid item xs={12} md={2}>
-          <FormControl fullWidth>
-            <Typography>Start Time</Typography>
-            <TextField
-              type="time"
-              name="start_time"
-              value={activity.start_time}
-              onChange={handleChange}
-              // className={classes.input}
-            />
-          </FormControl>
-        </Grid>
-        <Grid item xs={12} md={2}>
-          <FormControl fullWidth>
-            <Typography>End Time</Typography>
-            <TextField
-              type="time"
-              name="end_time"
-              value={activity.end_time}
-              onChange={handleChange}
-              // className={classes.input}
-            />
-          </FormControl>
-        </Grid>
-      </Grid>
-      {overlapError && (
-        <Typography variant="body2">
-          There is an overlap with another activity on this day and time.
-        </Typography>
-      )}
-
-      <br />
-      <FormControl>
-        <Button type="submit" variant="contained" color="primary">
-          Add Activity
-        </Button>
-      </FormControl>
-    </form>
-  );
-}
 export default WeekPage;
